+++ lib/tapioca/dsl/compilers/mixed_in_class_attributes.rb
@@ -67,7 +67,7 @@ module Tapioca
         def self.gather_constants
           # Select all non-anonymous modules that have overridden Module.included
           all_modules.select do |mod|
-            !mod.is_a?(Class) && name_of(mod) && Tapioca::Reflection.method_of(mod, :included).owner != Module
+            !mod.is_a?(Class) && name_of(mod) && Runtime::Reflection.method_of(mod, :included).owner != Module
           end
         end
       end
