+++ lib/tapioca/gem/listeners/sorbet_signatures.rb
@@ -7,7 +7,7 @@ module Tapioca
       class SorbetSignatures < Base
         extend T::Sig
 
-        include Reflection
+        include Runtime::Reflection
         include RBIHelper
 
         TYPE_PARAMETER_MATCHER = /T\.type_parameter\(:?([[:word:]]+)\)/
