+++ lib/tapioca/gem/listeners/subconstants.rb
@@ -7,7 +7,7 @@ module Tapioca
       class Subconstants < Base
         extend T::Sig
 
-        include Reflection
+        include Runtime::Reflection
 
         private
