+++ lib/tapioca/sorbet_ext/name_patch.rb
@@ -3,7 +3,7 @@
 
 # We need sorbet to compile the signature for `qualified_name_of` before applying
 # the patch to avoid an infinite loop.
-T::Utils.signature_for_method(::Tapioca::Reflection.method(:qualified_name_of))
+T::Utils.signature_for_method(::Tapioca::Runtime::Reflection.method(:qualified_name_of))
 
 module T
   module Types
@@ -12,7 +12,7 @@ module T
         def name
           # Sorbet memoizes this method into the `@name` instance variable but
           # doing so means that types get memoized before this patch is applied
-          ::Tapioca::Reflection.qualified_name_of(@raw_type)
+          ::Tapioca::Runtime::Reflection.qualified_name_of(@raw_type)
         end
       end
